rename: domain 패키지 명을 entity로 수정
