+++ src/test/java/com/devleoh/glomo/member/domain/MemberTest.java
@@ -1,5 +1,6 @@
 package com.devleoh.glomo.member.domain;
 
+import com.devleoh.glomo.member.entity.Member;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
