+++ src/test/java/com/devleoh/glomo/member/repository/MemberRepositoryTest.java
@@ -1,6 +1,6 @@
 package com.devleoh.glomo.member.repository;
 
-import com.devleoh.glomo.member.domain.Member;
+import com.devleoh.glomo.member.entity.Member;
 import com.devleoh.glomo.util.SHA256;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
